Move mapMenu into map and add a test for start Fixes #28
