+++ cliObjects/gameCli.py
@@ -1,5 +1,4 @@
 import _globals
-import random
 from dataFactory import dataFactory
 from cliObjects.menuFunctions import menuFunctions
 from dataObjects.enums.gameStates import GameStates
@@ -20,7 +19,7 @@ class GameCli:
         sys.exit()
     
     def mainMenu(self):
-        options : list[tuple[str, Callable]] = [("Fight", self.standaloneFight),("Random Map", self.mapMenu), ("Premade Map", self.pickPremadeMap), ("Quit", self.quit)]
+        options : list[tuple[str, Callable]] = [("Fight", self.standaloneFight),("Random Map", self.map.mapMenu), ("Premade Map", self.pickPremadeMap), ("Quit", self.quit)]
         if _globals.debug:
             options.append(("Debug", self.debugMenu))
         while self.gameState == GameStates.RUNNING:
@@ -46,55 +45,8 @@ class GameCli:
         mapDir : str = 'data/maps'
         maps : list[str] = dataFactory.getFilesInDirectory(mapDir)
         map = menuFunctions.menuStr(maps)
-        self.map : Map = Map.fromJson(mapDir + '/' + map)
-        self.mapMenu()
-    
-    def mapMenu(self):
-        if _globals.debug:
-            print(self.map)
-        
-        self.map.start()
-        exitMenu = False
-        while exitMenu != True:
-            options : list[tuple[str, Callable]] = [("Back", lambda : 1 == 1)]
-            if self.map.status is self.map.status.COMPLETED or self.map.status is self.map.status.FAILED:
-                options.append(("Reset map", self.map.reset))
-            else:
-                options.append(("Advance", self.advanceMenu))
-            exitMenu : bool = menuFunctions.menuCallable(options)
-            
-    def advanceMenu(self):
-        while self.map.status is not self.map.status.COMPLETED and self.map.status is not self.map.status.FAILED:
-            options : dict[int, str] = dict({
-                0 : "Back",
-                1 : "Status"
-            })
-            self.showCurrentNode()
-            offset : int = len(options)
-            self.addForwardLinks(options)
-            
-            if len(options) == offset:
-                return
-
-            pickedOption : int = menuFunctions.menuInt(options)
-            if pickedOption == 0:
-                return
-            elif pickedOption == 1:
-                _globals.player.status
-            else:
-                nodeIndex : int = pickedOption - offset
-                self.map.advance(self.map.activeNode.forelinks[nodeIndex])
-                if _globals.player.party.isDefeated():
-                    self.map.fail()
-                    print('Game over !')
-    
-    def showCurrentNode(self):
-        index = self.map.activeNode.columnIndex + 1
-        print(f"Current Node : {self.map.activeNode}. At {index}/{len(self.map.nodes)}")
-                    
-    def addForwardLinks(self, options):
-        for node in self.map.activeNode.forelinks:
-            options[len(options)] = str(node)
+        map : Map = Map.fromJson(mapDir + '/' + map)
+        map.mapMenu()                    
 
     def createPlayer(self):
         _globals.player = Player('test')
